Put a system wide status bar at the bottom.  It contains a general purpose progress bar and a "currently open file/folder" label.  Made these tools globally available from the SIPCreator.
Refactored the "parent" variable in all 3 tasks to be "creator"

CRules are now saved in the SIP file.

MinimalMetadata now puts out valid XML under all conditions.

I use the LABEL attribute in the structMap:div's to associate a DIV with a particular ZIP entry.  structMap:div:ID is now unused.

ConversionRules is now easily convertible to XML.

Some configuration stuff changed - the name on the MinimalMetadata class (in classList) and I added the JAF JAR file to the lib directory.