ConversionRules no longer has anything to do with Observable or Observer objects.
ConversionRules attribute are now immutable.  ie - you cannot change the name of a relationship or the nodetype of any template or the alias of a namespace.  These fields are now final and must be preset upon instantiation.

The ConversionRulesTask now allows for the user to add, delete, and edit namespaces and datastream/object template attributes.